Stop swallowing vendor business sync failures
