+++ controllers/vendorOnboarding.controller.js
@@ -12,6 +12,7 @@ const {
   applyVendorBusinessProfileFields,
   applyVendorDraftField,
 } = require('../utils/vendorOnboardingProfileFields');
+const { syncBusinessFromOnboarding } = require('../utils/syncBusinessFromOnboarding');
 
 /* =====================================================
    COMMON VALIDATION HELPERS
@@ -572,9 +573,6 @@ exports.updateBusinessProfile = async (req, res) => {
     const userId = req.user._id;
     const payload = req.body;
 
-    // Load models
-    const Business = require('../models/Business');
-    const Subscription = require('../models/Subscription');
     const VendorOnboarding = require('../models/VendorOnboardingStage1');
 
     // 1️⃣ Check existing onboarding
@@ -594,85 +592,22 @@ exports.updateBusinessProfile = async (req, res) => {
       isVendorProfileReadyForTrustBadgeVerification(onboarding);
     await onboarding.save();
 
-    // ========== SIMPLE BUSINESS SYNC ==========
+    // ========== REQUIRED BUSINESS SYNC ==========
     try {
-      // Get active subscription
-      const subscription = await Subscription.findOne({ 
-        userId, 
-        status: 'active' 
-      }).sort({ createdAt: -1 });
-
-      // Find or create business
-      let business = await Business.findOne({ owner: userId });
-
-      // ONLY THESE FIELDS - SIMPLE & CLEAN
-      const businessData = {
-        businessName: onboarding.businessName,
-        description: onboarding.businessBio,
-        logo: onboarding.businessProfileImage?.url,
-        coverImage: onboarding.featureBanner?.url,
-        email: onboarding.businessEmail || onboarding.secondaryBusinessEmail,
-        phone: onboarding.businessPhone || onboarding.primaryPhone,
-        listingType: onboarding.businessType || 'product',
-        points: onboarding.totalVerificationPoints || 0,
-        badge: onboarding.badge || null,
-        
-        // Subscription reference (important for limits)
-        subscriptionId: subscription?._id || null,
-        subscriptionPlanId: subscription?.subscriptionPlanId || null,
-        subscriptionStatus: subscription?.status || 'inactive',
-      };
-
-      if (!business) {
-        // Create new business
-        business = new Business({
-          owner: userId,
-          ...businessData,
-          isApproved: false,
-          isActive: true,
-          usage: {
-            totalProducts: 0,
-            totalServices: 0,
-            totalFoods: 0,
-            totalImages: 0,
-          },
-          products: [],
-          services: [],
-          foods: [],
-        });
-      } else {
-        // Update existing business
-        business.businessName = businessData.businessName;
-        business.description = businessData.description;
-        business.logo = businessData.logo;
-        business.coverImage = businessData.coverImage;
-        business.email = businessData.email;
-        business.phone = businessData.phone;
-        business.listingType = businessData.listingType;
-        business.points = businessData.points;
-        business.badge = businessData.badge;
-        business.subscriptionId = businessData.subscriptionId;
-        business.subscriptionPlanId = businessData.subscriptionPlanId;
-        business.subscriptionStatus = businessData.subscriptionStatus;
-        
-        // CRITICAL: Remove location field to prevent geo index error
-        if (business.location) {
-          business.location = undefined;
-        }
-      }
-
-      await business.save();
-
-      // Link onboarding <-> business for direct lookups in admin flows
-      if (!onboarding.businessId || onboarding.businessId.toString() !== business._id.toString()) {
-        onboarding.businessId = business._id;
-        await onboarding.save();
-      }
-
-      console.log(`✅ Business data saved for user ${userId}`);
-
+      const Business = require('../models/Business');
+      const Subscription = require('../models/Subscription');
+      await syncBusinessFromOnboarding({
+        userId,
+        onboarding,
+        Business,
+        Subscription,
+      });
     } catch (businessError) {
-      console.log('⚠️ Business sync issue:', businessError.message);
+      console.error('Business sync failed:', businessError.message);
+      return res.status(500).json({
+        success: false,
+        message: 'Failed to sync business profile data',
+      });
     }
 
     // Non-blocking: notify admin once when vendor completes profile + docs
